[GDPR Extractor] Consider entity permissions on actions #3443
