+++ bundles/AdminBundle/Resources/public/css/admin.css
@@ -2012,4 +2012,10 @@ ul.leaflet-draw-actions.leaflet-draw-actions-top.leaflet-draw-actions-bottom li:
     width: 100%;
 }
 
+.inactive_actioncolumn {
+    filter: Alpha(opacity=70) !important;
+    opacity: 0.7 !important;
+}
+
+
 /** END workflow engine **/
