+++ bundles/CoreBundle/Resources/translations/en.json
@@ -828,5 +828,6 @@
   "no_thumbnail_selected": "No thumbnail selected",
   "list_thumbnail_in_download_section_on_image_detail_view": "List as option in download section on image detail view",
   "there_are_more_items": "There are more items than displayed",
-  "unit_width": "Width Unit"
+  "unit_width": "Width Unit",
+  "permission_missing": "You need the '%s' permission to perform this action"
 }
